Added buffer allocator template parameter
Added the allocator of the buffer structure as a template parameter,
with default value equal to the type.hpp allocator of the value type.

Also refactored start and finish to be private and added functions for
accessing them.

Also removed redundant buffer methods construct_at and destruct_at.
Reason: the buffer only manages the raw bytes, not object
creation/deletion.